project renaming, documentation update, test dockerfile
