+++ .gitignore
@@ -161,5 +161,5 @@ cython_debug/
 #  and can be added to the global gitignore or merged into this file.  For a more nuclear
 #  option (not recommended) you can uncomment the following to ignore the entire idea folder.
 #.idea/
-toloka2anime/data/titles.ini
-toloka2anime/data/app.ini
+toloka2MediaServer/data/titles.ini
+toloka2MediaServer/data/app.ini
