+++ crontab
@@ -1,0 +1,1 @@
+$CRON_SCHEDULE root python3 -m toloka2MediaServer
