+++ toloka2MediaServer/data/app-example.ini
@@ -1,0 +1,32 @@
+[Python]
+# NOTSET
+# DEBUG
+# INFO
+# WARNING
+# ERROR
+# CRITICAL
+logging = INFO
+
+[Transmission]
+username = Імя користувача
+password = Пароль
+port = 9091
+host = localhost
+protocol = http
+rpc = /transmission/rpc
+category = sonarr
+
+[qBittorrent]
+username = Імя користувача
+password = Пароль
+port = 8080
+host = 192.168.40.22
+protocol = http
+tag = toloka
+category = sonarr
+
+[Toloka]
+username = 
+password = 
+client = qBittorrent
+default_download_dir = /media/HDD/Jellyfin/Anime
