+++ toloka2MediaServer/data/titles-example.ini
@@ -1,0 +1,169 @@
+[ArknightsTouin]
+episode_number = 2
+season_number = 02
+ext_name = .mkv
+torrent_name = "Arknights: Touin Kiro (2022)"
+download_dir = 
+publishdate = 2024-05-23
+release_group = InariDuB
+meta = [WEBRip-1080p][UK+JA][Ukr Sub]
+hash = 97e3023362ebb41263f3266ac3a72cc56eda0885
+adjusted_episode_number = -8
+guid = "t678205"
+
+[ArknightsReimei]
+episode_number = 0
+season_number = 01
+ext_name = .mkv
+torrent_name = "Arknights: Reimei Zensou (2022)"
+download_dir = 
+publishdate = 2023-03-06
+release_group = InariDuB
+meta = [WEBRip-1080p][UK+JA][Ukr Sub]
+hash = 4d298e6380dd98fcbf526a8fdb8f98b48e640700
+adjusted_episode_number = 0
+guid = "t666176"
+
+[DemonSlayer]
+episode_number = 3
+season_number = 05
+ext_name = .mkv
+torrent_name = "Demon Slayer - Kimetsu no Yaiba (2019)"
+download_dir = 
+publishdate = 2024-05-20
+release_group = Amanogawa
+meta = [WEBRip-1080p][UK+JA][Ukr Sub]
+hash = 8dcb775d10cbe2c7f78052d7a0beb5053a2a801b
+adjusted_episode_number = 0
+guid = "t678861"
+
+[WindBreaker]
+episode_number = 1
+season_number = 01
+ext_name = .mkv
+torrent_name = "Wind Breaker (2024)"
+download_dir = 
+publishdate = 2024-05-24
+release_group = Amanogawa
+meta = [WEBRip-1080p][UK+JA][Ukr Sub]
+hash = a247478627c6d5f66209b244d36e1fecbeee9bde
+adjusted_episode_number = 0
+guid = "t678025"
+
+[Unnamed]
+episode_number = 1
+season_number = 01
+ext_name = .mkv
+torrent_name = "Unnamed Memory (2024)"
+download_dir = 
+publishdate = 2024-05-22
+release_group = GlassMoon
+meta = [WEBRip-1080p][UK+JA][Ukr Sub]
+hash = 1e877bebd678c8b2eb5f7a711bf87e2350555d0e
+adjusted_episode_number = 0
+guid = "t677935"
+
+[Kaijuu]
+episode_number = 3
+season_number = 01
+ext_name = .mkv
+torrent_name = "Kaiju No. 8 (2024)"
+download_dir = 
+publishdate = 2024-05-26
+release_group = Amanogawa
+meta = [WEBRip-1080p][UK+JA][Ukr Sub]
+hash = 1243da5beedce34dabbe23aa68aff31f30253d7d
+adjusted_episode_number = 0
+guid = "t678236"
+
+[lv2]
+episode_number = 2
+season_number = 01
+ext_name = .mkv
+torrent_name = "Chillin' in Another World With Level 2 Super Cheat Powers (2024)"
+download_dir = 
+publishdate = 2024-05-26
+release_group = fanat22012
+meta = [WEBRip-1080p][UK+JA][Ukr Sub]
+hash = 2712f78e110bb9efb86a395a70d45a202c8c0c7e
+adjusted_episode_number = 0
+guid = "t678112"
+
+[Fable]
+episode_number = 2
+season_number = 01
+ext_name = .mkv
+torrent_name = "The Fable (2024)"
+download_dir = 
+publishdate = 2024-05-26
+release_group = GlassMoon
+meta = [WEBRip-1080p][UK+JA][Ukr Sub]
+hash = dd851b9e600a4914a7e187a61c9ffc9fbe26dd5d
+adjusted_episode_number = 0
+guid = "t677903"
+
+[MushokuTensei]
+episode_number = 5
+season_number = 02
+ext_name = .mkv
+torrent_name = "Mushoku Tensei - Jobless Reincarnation (2021)"
+download_dir = 
+publishdate = 2024-05-13
+release_group = Amanogawa
+meta = [WEBRip-1080p][UK+JA][Ukr Sub]
+hash = b6a560a141a69a414918a55d975bc48400c83800
+adjusted_episode_number = 0
+guid = "t678115"
+
+[KonoSuba]
+episode_number = 3
+season_number = 03
+ext_name = .mkv
+torrent_name = "KonoSuba – God’s blessing on this wonderful world!! (2016)"
+download_dir = 
+publishdate = 2024-05-23
+release_group = Amanogawa
+meta = [WEBRip-1080p][UK+JA][Ukr Sub]
+hash = 270fda72928873a47d0203e941a8b8c648d8d47e
+adjusted_episode_number = 0
+guid = "t678172"
+
+[WiseWolf]
+episode_number = 3
+season_number = 01
+ext_name = .mkv
+torrent_name = "Spice and Wolf - MERCHANT MEETS THE WISE WOLF (2024)"
+download_dir = 
+publishdate = 2024-05-21
+release_group = Amanogawa
+meta = [WEBRip-1080p][UK+JA][Ukr Sub]
+hash = 31752110d9b36d50425ccbc9b1310650f223bef7
+adjusted_episode_number = 0
+guid = "t677952"
+
+[Slime]
+episode_number = 2
+season_number = 03
+ext_name = .mkv
+torrent_name = "That Time I Got Reincarnated as a Slime (2018)"
+download_dir = 
+publishdate = 2024-05-26
+release_group = Kioto_anime
+meta = [WEBRip-1080p][UK][Ukr Sub]
+hash = 7b6dd63c229c7dfd50c09bca0831b9fd13783d5c
+adjusted_episode_number = 0
+guid = "t678077"
+
+[Tsukimichi]
+episode_number = 2
+season_number = 02
+ext_name = .mkv
+torrent_name = "Tsukimichi -Moonlit Fantasy- (2021)"
+download_dir = 
+publishdate = 2024-05-21
+release_group = FanVoxUA
+meta = [WEBRip-1080p][UK][Ukr Sub]
+hash = 8bcb2b32b4885e6c4a03f909486a03f26a4c9a62
+adjusted_episode_number = 0
+guid = "t675888"
+
